// src/app/models/reports/detalle-unidades-acumulado.ts
export interface IDetalleUnidadesAcumulado {
  departamento: string;
  descripcion: string;
  IdAutoLinea: number;
  autoLinea: string;
  enero: number;
  eneroPerc: number;
  febrero: number;
  febreroPerc: number;
  marzo: number;
  marzoPerc: number;
  abril: number;
  abrilPerc: number;
  mayo: number;
  mayoPerc: number;
  junio: number;
  junioPerc: number;
  julio: number;
  julioPerc: number;
  agosto: number;
  agostoPerc: number;
  septiembre: number;
  septiembrePerc: number;
  octubre: number;
  octubrePerc: number;
  noviembre: number;
  noviembrePerc: number;
  diciembre: number;
  diciembrePerc: number;
  totalAnual: number;
  totalAnualPerc: number;
  idOrden: number;
  idEstadoResultadosI: number;

}